Use delay from avr-libc.
delay() is from the Arduino library, and assumes that init() has been called,
initializing TIMER0. Without that, the call just hangs. _delay_ms() is a macro
that just spins based on the clock, which is just fine for my purposes.

This library also uses millis(), which also depends on this clock, for doing
DHCP timeout (and probably other things). So those timeouts don't work until
init() is called.